requirements.txt
# Core Django
Django==5.2.6
djangorestframework==3.15.0

# Real-time features
channels==4.1.0
channels-redis==4.2.0
asgiref==3.8.1
daphne==4.1.0

# Background tasks & Celery
celery==5.3.6
redis==5.0.1
django-celery-beat==2.7.0
django-celery-results==2.5.0

# API & CORS
django-cors-headers==4.4.0
drf-yasg==1.21.7
coreapi==2.3.3

# Static files
whitenoise==6.7.0

# Utilities
requests==2.31.0
pytz==2024.1
python-dotenv==1.0.1
sqlparse==0.5.0

# File processing & AI
markdown==3.6
python-docx==1.1.0
PyPDF2==3.0.1
pillow==10.4.0
pandas==2.2.2
beautifulsoup4==4.12.3
python-magic==0.4.27
python-magic-bin==0.4.14; sys_platform == 'win32'

# Database (if using PostgreSQL)
psycopg2-binary==2.9.9

# Development
debugpy==1.8.1